+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add implementation function for `dict.pop` (#517)
+- Remove `WeakExtension` (#517)
 - Initial support for variable-length heterogeneous sequences
   (required for PEP 646). More precise types are now inferred
   for heterogeneous sequences containing variable-length